Desktop/test/requirements.txt
asgiref==3.2.5
certifi==2019.11.28
chardet==3.0.4
Django==3.0.4
django-ckeditor==5.9.0
django-js-asset==1.2.2
django-reversion==3.0.7
djangorestframework==3.11.0
facebook-sdk==3.1.0
idna==2.9
Pillow==7.0.0
pytz==2019.3
requests==2.23.0
sqlparse==0.3.1
urllib3==1.25.8
